renamed bzx to ooki + changed polygon and bsc affreses + added staking v2
